ref: move EditRadio and EditCheckbox to own folder
so we can add stories later on to prevent regression
