+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/EditFieldDrawer/edit-fieldtype/EditCheckbox/index.ts
@@ -1,0 +1,1 @@
+export * from './EditCheckbox'
